refactor: consolidate tag components and update imports for better organization
