+++ src/components/Modal/index.tsx
@@ -21,7 +21,6 @@ import ReportProfile from './_ReportProfile';
 import SearchedUsersCard from './_SearchedUsersCard';
 import SearchInputCard from './_SearchInputCard';
 import SessionExpired from './_SessionExpired';
-import Tag from './_Tag';
 import TagCreatePost from './_TagCreatePost';
 
 export const Modal = {
@@ -48,7 +47,6 @@ export const Modal = {
   SearchedUsersCard,
   SearchInputCard,
   SessionExpired,
-  Tag,
   TagCreatePost
 };
