--- src/components/Post/Tags/index.tsx
@@ -1,9 +1,0 @@
-import { Standard } from './_Standard';
-import { LargeView } from './_LargeView';
-
-export const Tags = {
-  Standard,
-  LargeView
-};
-
-export default Tags;
