+++ src/components/Tags/components/index.tsx
@@ -1,0 +1,13 @@
+import InputTag from './_InputTag';
+import InputTagStandard from './_InputTagStandard';
+import InputTag2 from './_InputTag2';
+import Tags2 from './_Tags2';
+
+export const Components = {
+  InputTag,
+  InputTagStandard,
+  InputTag2,
+  Tags2
+};
+
+export default Components;
