+++ src/components/Tags/index.tsx
@@ -1,0 +1,13 @@
+import { PostStandard } from './_PostStandard';
+import { PostLargeView } from './_PostLargeView';
+import { PostBottomSheet } from './_PostBottomSheet';
+import { PostModal } from './_PostModal';
+
+export const Tags = {
+  PostStandard,
+  PostLargeView,
+  PostBottomSheet,
+  PostModal
+};
+
+export default Tags;
