+++ src/components/Tags/utils/index.tsx
@@ -1,0 +1,12 @@
+import { handleAddTag, handleDeleteTag } from './_TagsCommonFunctions';
+import { useUtilsTag } from './_TagsUtils2';
+import TagsUtils from './_TagsUtils';
+
+export const Utils = {
+  handleAddTag,
+  handleDeleteTag,
+  useUtilsTag,
+  TagsUtils
+};
+
+export default Utils;
